rename app install –> vpp install modal
