--- frontend/components/ActivityDetails/InstallDetails/AppInstallDetails/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppInstallDetails";
