+++ frontend/components/ActivityDetails/InstallDetails/VppInstallDetailsModal/_styles.scss
@@ -1,4 +1,4 @@
-.app-install-details {
+.vpp-install-details-modal {
   overflow-wrap: anywhere; // Prevent long software name overflow
 
   &__modal-content {
