+++ frontend/components/ActivityDetails/InstallDetails/VppInstallDetailsModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./VppInstallDetailsModal";
